+++ packages/integrations/vue-devtools/functions/defineDevToolsSettings.ts
