+++ packages/integrations/vue-devtools/utils/config.ts
@@ -1,4 +1,4 @@
-import { version } from '../package.json'
+import { version } from '../../package.json'
 
 export const id = `@exposition/vue-devtools/${version}`
