+++ packages/integrations/vue-devtools/utils/logs.ts
